Extract Formatter class from TextUi class
